// src/app/core/interfaces/accredited-garages-search-result.interface.ts
export interface AccreditedGaragesSearchResult {
  id?: number;
  name?: string;
  address?: string;
  phoneNumber?: string;
  email?: string;
  distance?: string;
  latitude?: number;
  longitude?: number;
  technicalManager?: string;
  attendance?: string;
  accessibility?: string;
  regionalCouncil?: string;
  typeOfProvider?: string;
  notes?: string;
}